fix: Add deleteOrphans flag.

+++ packages/orm/src/relations/OneToManyCollection.ts
@@ -109,14 +109,15 @@ export class OneToManyCollection<T extends Entity, U extends Entity>
     return this.loaded;
   }
 
-  set(values: U[]): void {
+  set(values: U[], opts: { deleteOrphans?: boolean } = {}): void {
     ensureNotDeleted(this.entity);
     if (this.loaded === undefined) {
       throw new Error("set was called when not loaded");
     }
 
     // If we're changing `a1.books = [b1, b2]` to `a1.books = [b2]`, then implicitly delete the old book
-    if (this.isCascadeDelete) {
+    const { deleteOrphans = false } = opts;
+    if (this.isCascadeDelete && deleteOrphans) {
       const implicitlyDeleted = this.loaded.filter((e) => !values.includes(e));
       implicitlyDeleted.forEach((e) => this.entity.em.delete(e));
       // Keep the implicitlyDeleted values for `getWithDeleted` to return
